residual_masking
added making based on the residuals (span of 0.25A either side of the wavelength above tolerance). Also added a few smaller changes such as centering the profile in the velocity range and added a penalty functon on the outer 3 point of the profile, and making the masking array a csv file so it can be easily added to.